Update Dart compiler and regenerate machine outputs

+++ tests/machine/x/dart/list_assign.dart
@@ -1,6 +1,6 @@
 var nums = [1, 2];
 
 void main() {
-  nums[1] = 3;
+  nums[1] = (3 as int);
   print(nums[1]);
 }
