+++ tests/machine/x/dart/map_assign.dart
@@ -1,6 +1,6 @@
 var scores = {'alice': 1};
 
 void main() {
-  scores['bob'] = 2;
+  scores['bob'] = (2 as int);
   print((scores as Map)['bob']);
 }
