+++ tests/machine/x/dart/while_loop.dart
@@ -3,6 +3,6 @@ var i = 0;
 void main() {
   while (i < 3) {
     print(i);
-    i = i + 1;
+    i = (i + 1 as int);
   }
 }
